pyproject.toml
[tool.poetry]
name = "test-reksoft"
version = "0.1.0"
description = ""
authors = ["basbky <markbasbky@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.5"
black = "^22.12.0"
pre-commit = "^3.0.0"
pytest = "^7.2.1"
pytest-django = "^4.5.2"
flake8 = "^6.0.0"
python-dotenv = "^0.21.1"
django-mathfilters = "^1.0.0"
isort = "^5.11.4"
factory-boy = "^3.2.1"
django-debug-toolbar = "^3.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
skip-string-normalization = true
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''